[design] 03/10 소설 검토 표 생성 #319
